requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-guardian==2.4.0
djangorestframework==3.14.0
djangorestframework-guardian==0.3.0
